// old/src/app/dataTransferObjects/monitor/eshopwms/event.search.general.dto.ts
export interface IEventSearchGeneralDto {
  idRegistro?: string;
  idIntegrador?: string;
  codigo?: string;
  keyTipo?: string;
  tipo?: string;
  idShopify?: string;
  idShopifyBusiness?: string;
  keyTipoEventCode?: string;
  tipoEventCode?: string;
  keyEstado?: string;
  estado?: string;
  errorMessage?: string;
  jsonResponse?: string;
  createDateUtc?: string;
  origenDate?: string;
  envioDate?: string;
  errorCode?: string;
  errorCodeBusiness?: string;
  dCreateDateUtcDate?: Date;
  dOrigenDate?: Date;
  dEnvioDate?: Date;
}